#239 Remove duplicated code
Remove Diagram ExportAction copy and replace it with correct call to
ExportAction.
